+++ amr-wind/utilities/PlaneAveraging.cpp
@@ -288,7 +288,7 @@ Real PlaneAveraging::eval_line_average(Real x, int comp) const
     int ind = 0;
 
     if (x > xlo + 0.5 * dx) {
-        ind = floor((x - xlo) / dx - 0.5);
+        ind = static_cast<int>(floor((x - xlo) / dx - 0.5));
         const Real x1 = xlo + (ind + 0.5) * dx;
         c = (x - x1) / dx;
     }
